lostfound/requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.0
bootstrap4==0.1.0
Django==2.2.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.0.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-filters==0.2.1
django-tempus-dominus==5.1.2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.1.0
pkg-resources==0.0.0
pylint==2.3.1
pytz==2019.2
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
